Delete unused callback
