+++ gradle.properties
@@ -54,7 +54,6 @@ jettyVersion=12.0.5
 
 # rest_tools
 swaggerRequestValidatorVersion=2.40.0
-vertxVersion=4.5.1
 
 # serialization
 jacksonVersion=2.16.1
